Separate errand and auto errand integration tests
