+++ integration/auto_errand_job_test.go
@@ -7,31 +7,17 @@ import (
 	. "github.com/onsi/gomega"
 
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	ejv1 "code.cloudfoundry.org/quarks-job/pkg/kube/apis/extendedjob/v1alpha1"
-	"code.cloudfoundry.org/quarks-utils/pkg/pointers"
 	"code.cloudfoundry.org/quarks-utils/testing/machine"
 )
 
-var _ = Describe("ExtendedJob", func() {
-	jobOwnerRef := func(eJob ejv1.ExtendedJob) metav1.OwnerReference {
-		return metav1.OwnerReference{
-			APIVersion:         "fissile.cloudfoundry.org/v1alpha1",
-			Kind:               "ExtendedJob",
-			Name:               eJob.Name,
-			UID:                eJob.UID,
-			Controller:         pointers.Bool(true),
-			BlockOwnerDeletion: pointers.Bool(true),
-		}
-	}
-
+var _ = Describe("AutoErrandJob", func() {
 	AfterEach(func() {
 		env.FlushLog()
 	})
 
 	Context("when using an AutoErrandJob", func() {
-
 		var (
 			ej        ejv1.ExtendedJob
 			tearDowns []machine.TearDownFunc
@@ -314,68 +300,6 @@ var _ = Describe("ExtendedJob", func() {
 					Expect(err).ToNot(HaveOccurred())
 				})
 			})
-
-		})
-
-	})
-
-	Context("when using manually triggered ErrandJob", func() {
-		It("does not start a job without Run being set to now", func() {
-			ej := env.ErrandExtendedJob("extendedjob")
-			ej.Spec.Trigger.Strategy = ejv1.TriggerManual
-			_, tearDown, err := env.CreateExtendedJob(env.Namespace, ej)
-			Expect(err).NotTo(HaveOccurred())
-			defer func(tdf machine.TearDownFunc) { Expect(tdf()).To(Succeed()) }(tearDown)
-
-			exists, err := env.WaitForJobExists(env.Namespace, fmt.Sprintf("%s=true", ejv1.LabelExtendedJob))
-			Expect(err).NotTo(HaveOccurred())
-			Expect(exists).To(BeFalse())
-
-			latest, err := env.GetExtendedJob(env.Namespace, ej.Name)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(latest.Spec.Trigger.Strategy).To(Equal(ejv1.TriggerManual))
-
-			err = env.UpdateExtendedJob(env.Namespace, *latest)
-			Expect(err).NotTo(HaveOccurred())
-
-			exists, err = env.WaitForJobExists(env.Namespace, fmt.Sprintf("%s=true", ejv1.LabelExtendedJob))
-			Expect(err).NotTo(HaveOccurred())
-			Expect(exists).To(BeFalse())
-		})
-
-		It("starts a job when creating extended job with now", func() {
-			ej := env.ErrandExtendedJob("extendedjob")
-			_, tearDown, err := env.CreateExtendedJob(env.Namespace, ej)
-			Expect(err).NotTo(HaveOccurred())
-			defer func(tdf machine.TearDownFunc) { Expect(tdf()).To(Succeed()) }(tearDown)
-
-			jobs, err := env.CollectJobs(env.Namespace, fmt.Sprintf("%s=true", ejv1.LabelExtendedJob), 1)
-			Expect(err).NotTo(HaveOccurred(), "error waiting for jobs from extendedjob")
-			Expect(jobs).To(HaveLen(1))
-		})
-
-		It("starts a job when updating extended job to now", func() {
-			ej := env.ErrandExtendedJob("extendedjob")
-			ej.Spec.Trigger.Strategy = ejv1.TriggerManual
-			_, tearDown, err := env.CreateExtendedJob(env.Namespace, ej)
-			Expect(err).NotTo(HaveOccurred())
-			defer func(tdf machine.TearDownFunc) { Expect(tdf()).To(Succeed()) }(tearDown)
-
-			latest, err := env.GetExtendedJob(env.Namespace, ej.Name)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(latest.Spec.Trigger.Strategy).To(Equal(ejv1.TriggerManual))
-
-			latest.Spec.Trigger.Strategy = ejv1.TriggerNow
-			err = env.UpdateExtendedJob(env.Namespace, *latest)
-			Expect(err).NotTo(HaveOccurred())
-
-			jobs, err := env.CollectJobs(env.Namespace, fmt.Sprintf("%s=true", ejv1.LabelExtendedJob), 1)
-			Expect(err).NotTo(HaveOccurred(), "error waiting for jobs from extendedjob")
-			Expect(jobs).To(HaveLen(1))
-
-			Expect(jobs[0].GetOwnerReferences()).Should(ContainElement(jobOwnerRef(*latest)))
-
 		})
 	})
-
 })
